Moved Environments into constants for consistency
